# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MindVault AI Personal Workspace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /gemini/conversations:
    get:
      operationId: listGeminiConversations
      tags: [gemini]
      summary: List all conversations
      responses:
        "200":
          description: List of conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiConversation"
    post:
      operationId: createGeminiConversation
      tags: [gemini]
      summary: Create a new conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateGeminiConversationBody"
      responses:
        "201":
          description: Created conversation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversation"
  /gemini/conversations/{id}:
    get:
      operationId: getGeminiConversation
      tags: [gemini]
      summary: Get conversation with messages
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Conversation with messages
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiConversationWithMessages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
    delete:
      operationId: deleteGeminiConversation
      tags: [gemini]
      summary: Delete a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeminiError"
  /gemini/conversations/{id}/messages:
    get:
      operationId: listGeminiMessages
      tags: [gemini]
      summary: List messages in a conversation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeminiMessage"
    post:
      operationId: sendGeminiMessage
      tags: [gemini]
      summary: Send a message and receive AI response (SSE stream)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendGeminiMessageBody"
      responses:
        "200":
          description: SSE stream of assistant response chunks
          content:
            text/event-stream: {}
  /gemini/classify:
    post:
      operationId: classifyContent
      tags: [gemini]
      summary: Classify content and auto-save as item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassifyContentBody"
      responses:
        "200":
          description: Classification result with optional saved item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassifyContentResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Search
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    # Auth schemas
...
    Item:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        folderId:
          type: ["integer", "null"]
        folderName:
          type: ["string", "null"]
        type:
          type: string
          enum: [note, file, reminder]
        title:
          type: string
        content:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        originalFilename:
          type: ["string", "null"]
        mimeType:
          type: ["string", "null"]
        fileSize:
          type: ["integer", "null"]
        fileData:
          type: ["string", "null"]
        reminderAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [active, archived, completed]
        aiCategory:
          type: ["string", "null"]
        aiTags:
          type: array
          items:
            type: string
        aiConfidence:
          type: ["number", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - type
        - title
        - status
        - aiTags
        - createdAt
        - updatedAt
...
    ClassifyContentBody:
      type: object
      properties:
        content:
          type: string
        conversationId:
          type: ["integer", "null"]
          description: Optional conversation context
      required:
        - content
    ClassifyContentResponse:
      type: object
      properties:
        type:
          type: string
          enum: [note, reminder, file, chat]
        title:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        cleanedContent:
          type: ["string", "null"]
        suggestedFolder:
          type: ["string", "null"]
        tags:
          type: array
          items:
            type: string
        confidence:
          type: number
        reminderAt:
          type: ["string", "null"]
        savedItem:
          nullable: true
          $ref: "#/components/schemas/Item"
        message:
          type: string
      required:
        - type
        - tags
        - confidence
        - message

    # Workspace stats
...
    GeminiConversation:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    GeminiMessage:
      type: object
      properties:
        id:
          type: integer
        conversationId:
          type: integer
        role:
          type: string
        content:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - conversationId
        - role
        - content
        - createdAt
    CreateGeminiConversationBody:
      type: object
      properties:
        title:
          type: string
      required:
        - title
    SendGeminiMessageBody:
      type: object
      properties:
        content:
          type: string
      required:
        - content
    GeminiConversationWithMessages:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
          format: date-time
        messages:
          type: array
          items:
            $ref: "#/components/schemas/GeminiMessage"
      required:
        - id
        - title
        - createdAt
        - messages
...
    GeminiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error