# lib/api-zod/src/generated/types/classifyContentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindVault AI Personal Workspace API
 * OpenAPI spec version: 0.1.0
 */
import type { ClassifyContentResponseType } from "./classifyContentResponseType";
import type { Item } from "./item";

export interface ClassifyContentResponse {
  type: ClassifyContentResponseType;
  /** @nullable */
  title?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  cleanedContent?: string | null;
  /** @nullable */
  suggestedFolder?: string | null;
  tags: string[];
  confidence: number;
  /** @nullable */
  reminderAt?: string | null;
  savedItem?: Item | null;
  message: string;
}