# lib/api-zod/src/generated/types/createGeminiConversationBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * MindVault AI Personal Workspace API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGeminiConversationBody {
  title: string;
}